fix(ui): medical mode was leaving a measured GPU machine on WASM
Two defects behind the same report ("mode med shows WASM, but clicking
Autoconfigure picks WebGPU").

1. The preset asked for fp16 on the GPU. fp16 needs BOTH a shader-f16
   adapter and a source hosting encoder-model.fp16.onnx; when either is
   missing the load raises QuantUnavailableError, App.jsx falls back to
   WASM and PERSISTS that flip, so the station quietly stops using its
   GPU. fp32 needs no adapter feature and every source shipping the
   <2 GB shards can serve it, so the preset now asks for that. int8 is
   not an option on the GPU: ORT has no WebGPU kernel for it.

2. The page-load autoconfigure treated "already measured" as "do
   nothing". A stored verdict is an answer we own and have to apply,
   because the live backend can have drifted away from it, and after
   defect 1 it always had: the persisted fallback left the machine on
   WASM while the verdict still said WebGPU, and nothing put the two
   back in agreement.

Re-applying a verdict must not fight the fallback that caused the drift
every visit, so a new persisted gpuWeightsUnservableSig records the
repo+precision that was proven unservable, and a stored WebGPU verdict
is skipped while it still matches. It self-invalidates when either half
changes, so hosting the missing files puts the GPU back in play with no
reset needed.

+++ app/ui/src/lib/medMode.js
@@ -51,6 +51,17 @@ export const MED_MODE_ALIASES = Object.freeze([
  *  - wasmEncoderQuant / webgpuEncoderQuant are set as a PAIR, not as one value
  *    for the current backend, because the on-load autoconfigure probe may still
  *    move the visitor between the two after this preset has been applied.
+ *  - webgpuEncoderQuant is fp32, NOT the smaller fp16, on purpose. fp16 needs
+ *    two things at once that a medical deployment often cannot promise: an
+ *    adapter exposing `shader-f16`, and a model source hosting
+ *    `encoder-model.fp16.onnx`. When either is missing the load raises
+ *    QuantUnavailableError, App.jsx falls the visitor back to WASM and PERSISTS
+ *    that flip, so the station quietly stops using its GPU. fp32 needs no
+ *    adapter feature and every repo that ships the <2 GB shards can serve it,
+ *    so it is the precision that actually works on a locked-down box. int8 is
+ *    NOT an option here despite being the WASM choice: ORT has no WebGPU kernel
+ *    for it (see WEBGPU_ENCODER_QUANTS in App.jsx), so asking for it on the GPU
+ *    would either be refused or silently execute on the CPU.
  *  - lang is forced to French: the lexicon, the dictation regexes and the model
  *    are all French, so an English UI would misdescribe the station.
  *  - autoCopyToClipboard is the one preset value that turns a default OFF into
@@ -69,7 +80,7 @@ export const MED_MODE_PRESET = Object.freeze({
   transcriptDisplayMode: 'dictation',
   autoCopyToClipboard: true,
   wasmEncoderQuant: 'int8',
-  webgpuEncoderQuant: 'fp16',
+  webgpuEncoderQuant: 'fp32',
   lang: 'fr',
 });
